Opened possibility to add specific algebraic structures
Created Groupoids, Semigroups and Monoids.

Added updates to operations and Algebra in order to adapt them to a specific algebraic structures environment (all new methods are protected and/or only have their return type changed to a subclass).

In order for all future changes to be more easily manipulated with, created BinaryOperationUtils, which contains method BinaryOperationUtils.isCommutative, BinaryOperationUtils.isAssociative and BinaryOperationUtils.findNeutralElement.